[SDPA] Migrate ring-attention op tests out of tt_dit (no unify)
Relocate ring_joint_attention / exp_ring_joint_attention op tests from
models/tt_dit/tests/unit into tests/ttnn/unit_tests/operations/sdpa, and
extract their shared helpers/configs/decorators into a single SDPA-owned
module (ring_joint_sdpa_test_common.py). Repoint the per-arch nightly
wrappers (t3000 / tg / blackhole) to import from the shared module.

Pure migration: no test logic changed. No pipeline's executed test set
changes - the relocated multi-card files are excluded from the single-card
post-commit sdpa glob via --ignore (run-set verified unchanged).
